hyperliquid-bot/requirements.txt
hyperliquid-python-sdk>=0.5.0
pandas>=2.1.0
pandas-ta>=0.3.14b
numpy>=1.25.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
sqlalchemy>=2.0.0
apscheduler>=3.10.0
loguru>=0.7.0
pyyaml>=6.0
python-dotenv>=1.0.0
httpx>=0.25.0
pydantic>=2.5.0
